+++ PoliNetworkBot_CSharp/Code/Utils/HtmlUtil.cs
@@ -8,7 +8,7 @@ namespace PoliNetworkBot_CSharp.Code.Utils
     internal class HtmlUtil
     {
         internal static List<HtmlNode> GetElementsByTagAndClassName(HtmlNode doc, string tag = "",
-            string className = "", int? limit = null)
+            string className = "", long? limit = null)
         {
             if (doc == null)
                 return null;
